Fixes
Fixed delete not working, changed bid and price back to Decimal field
because troll

+++ stocks/models.py
@@ -3,8 +3,8 @@ from django.db import models
 class Stock(models.Model):
     time= models.FloatField()
     symbol = models.CharField(max_length=100)
-    bid = models.FloatField()
-    ask = models.FloatField()
+    bid = models.DecimalField(max_digits=15, decimal_places=5)
+    ask = models.DecimalField(max_digits=15, decimal_places=5)
     exchange = models.IntegerField(default=1)
 
     class Meta:
